replace loose BaseQ with MinBaseQ + stricter distinct types
